move coverage utilities into subfolder
